rid of completion blinking and clean up code a bit
